2. Simple_LangChain_Application/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
langchain-openai
beautifulsoup4
faiss-cpu
streamlit